# src/IO.Swagger/Models/Livro.cs
/*
 * Livraria Virtual
 *
 * Manipulação de livros e suas dependências 
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class Livro : IEquatable<Livro>
    { 
        /// <summary>
        /// Gets or Sets Isbn
        /// </summary>
        [DataMember(Name="isbn")]
        public decimal? Isbn { get; set; }

        /// <summary>
        /// Gets or Sets Titulo
        /// </summary>
        [DataMember(Name="titulo")]
        public string Titulo { get; set; }

        /// <summary>
        /// Gets or Sets Autor
        /// </summary>
        [DataMember(Name="autor")]
        public Autor Autor { get; set; }

        /// <summary>
        /// Gets or Sets Genero
        /// </summary>
        [DataMember(Name="genero")]
        public Genero Genero { get; set; }

        /// <summary>
        /// Gets or Sets Data
        /// </summary>
        [DataMember(Name="data")]
        public DateTime? Data { get; set; }

        /// <summary>
        /// Gets or Sets Preco
        /// </summary>
        [DataMember(Name="preco")]
        public decimal? Preco { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Livro {\n");
            sb.Append("  Isbn: ").Append(Isbn).Append("\n");
            sb.Append("  Titulo: ").Append(Titulo).Append("\n");
            sb.Append("  Autor: ").Append(Autor).Append("\n");
            sb.Append("  Genero: ").Append(Genero).Append("\n");
            sb.Append("  Data: ").Append(Data).Append("\n");
            sb.Append("  Preco: ").Append(Preco).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Livro)obj);
        }

        /// <summary>
        /// Returns true if Livro instances are equal
        /// </summary>
        /// <param name="other">Instance of Livro to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Livro other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Isbn == other.Isbn ||
                    Isbn != null &&
                    Isbn.Equals(other.Isbn)
                ) && 
                (
                    Titulo == other.Titulo ||
                    Titulo != null &&
                    Titulo.Equals(other.Titulo)
                ) && 
                (
                    Autor == other.Autor ||
                    Autor != null &&
                    Autor.Equals(other.Autor)
                ) && 
                (
                    Genero == other.Genero ||
                    Genero != null &&
                    Genero.Equals(other.Genero)
                ) && 
                (
                    Data == other.Data ||
                    Data != null &&
                    Data.Equals(other.Data)
                ) && 
                (
                    Preco == other.Preco ||
                    Preco != null &&
                    Preco.Equals(other.Preco)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Isbn != null)
                    hashCode = hashCode * 59 + Isbn.GetHashCode();
                    if (Titulo != null)
                    hashCode = hashCode * 59 + Titulo.GetHashCode();
                    if (Autor != null)
                    hashCode = hashCode * 59 + Autor.GetHashCode();
                    if (Genero != null)
                    hashCode = hashCode * 59 + Genero.GetHashCode();
                    if (Data != null)
                    hashCode = hashCode * 59 + Data.GetHashCode();
                    if (Preco != null)
                    hashCode = hashCode * 59 + Preco.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Livro left, Livro right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Livro left, Livro right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}