# src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Manipulação de livros e suas dependências ",
    "version" : "v1",
    "title" : "Livraria Virtual"
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/IKokiri/Livrariav/v1",
  "schemes" : [ "https" ],
  "paths" : {
    "/livros" : {
      "get" : {
        "tags" : [ "livro" ],
        "summary" : "Busca livros do armazenamento",
        "operationId" : "getLivros",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Livro"
              }
            }
          },
          "400" : {
            "description" : "não encontrado"
          }
        }
      },
      "post" : {
        "tags" : [ "livro" ],
        "summary" : "Adiciona um novo livro",
        "operationId" : "addLivro",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "dados do livro que será adicionado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Livro"
          }
        } ],
        "responses" : {
          "405" : {
            "description" : "Dados inválidos"
          }
        }
      }
    },
    "/livros/autor/{autorId}/livros" : {
      "get" : {
        "tags" : [ "livro" ],
        "summary" : "Busca todos os livros de um autor",
        "description" : "faz a busca no banco dos livros por autor",
        "operationId" : "getLivrosByAutor",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "autorId",
          "in" : "path",
          "description" : "ID do autor",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Livro"
              }
            }
          },
          "400" : {
            "description" : "Id inválido"
          },
          "404" : {
            "description" : "Não Encontrado"
          }
        }
      }
    },
    "/livros/genero/{generoId}/livros" : {
      "get" : {
        "tags" : [ "livro" ],
        "summary" : "Busca todos os livros de um genero",
        "description" : "faz a busca no banco dos livros por genero",
        "operationId" : "getLivrosByGenero",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "generoId",
          "in" : "path",
          "description" : "ID do genero",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Livro"
              }
            }
          },
          "400" : {
            "description" : "Id inválido"
          },
          "404" : {
            "description" : "Não Encontrado"
          }
        }
      }
    },
    "/livros/{isbn}" : {
      "get" : {
        "tags" : [ "livro" ],
        "summary" : "Busca livro por isbn",
        "description" : "Retorna um unico livro",
        "operationId" : "getLivroByIsbn",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "isbn",
          "in" : "path",
          "description" : "identificador do livro de retorno",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "seccess",
            "schema" : {
              "$ref" : "#/definitions/Livro"
            }
          },
          "400" : {
            "description" : "ISBN inválido"
          },
          "404" : {
            "description" : "Livro não encontrado"
          }
        }
      },
      "put" : {
        "tags" : [ "livro" ],
        "summary" : "Atualização de um livro",
        "description" : "Efetua a alteração de um ou varios dados de um livro",
        "operationId" : "updateLivro",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "isbn",
          "in" : "path",
          "description" : "identificador do livro que será alterado",
          "required" : true,
          "type" : "number"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "dados do livro que serão alterados",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Livro"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Livro inválido"
          },
          "404" : {
            "description" : "Livro não encontrado"
          }
        }
      },
      "delete" : {
        "tags" : [ "livro" ],
        "summary" : "Remoção de um livro",
        "description" : "Remove um livro",
        "operationId" : "deleteLivro",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "isbn",
          "in" : "path",
          "description" : "identificador do livro que será removido",
          "required" : true,
          "type" : "number"
        } ],
        "responses" : {
          "400" : {
            "description" : "Livro inválido"
          },
          "404" : {
            "description" : "Livro não encontrado"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Livro" : {
      "type" : "object",
      "properties" : {
        "isbn" : {
          "type" : "number"
        },
        "titulo" : {
          "type" : "string"
        },
        "autor" : {
          "$ref" : "#/definitions/Autor"
        },
        "genero" : {
          "$ref" : "#/definitions/Genero"
        },
        "data" : {
          "type" : "string",
          "format" : "date-time"
        },
        "preco" : {
          "type" : "number"
        }
      },
      "xml" : {
        "name" : "Livro"
      },
      "example" : {
        "preco" : 5.962133916683182377482808078639209270477294921875,
        "data" : "2000-01-23T04:56:07.000+00:00",
        "isbn" : 0.80082819046101150206595775671303272247314453125,
        "genero" : {
          "genero" : "genero",
          "id" : 1
        },
        "titulo" : "titulo",
        "autor" : {
          "nome" : "nome",
          "id" : 6
        }
      }
    },
    "Autor" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "nome" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Autor"
      },
      "example" : {
        "nome" : "nome",
        "id" : 6
      }
    },
    "Genero" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "genero" : {
          "type" : "string"
        }
      },
      "xml" : {
        "name" : "Genero"
      },
      "example" : {
        "genero" : "genero",
        "id" : 1
      }
    },
...
  }
}